Extract API subpackage (#652)
* remove api folder

* remove api from docs

* Update tests/api/conftest.py
