--- doc/code/sf_api.rst
@@ -1,14 +1,0 @@
-sf.api
-======
-
-.. currentmodule:: strawberryfields.api
-
-.. warning::
-
-    Unless you are a Strawberry Fields developer, you likely do not need
-    to use these classes directly.
-
-.. automodapi:: strawberryfields.api
-    :no-heading:
-    :no-inheritance-diagram:
-    :include-all-objects:
