+++ strawberryfields/result.py
